Mali Bifrost DDK r8p0 KMD
Provenance:
 a0013c522 (collaborate/EAC/b_r8p0)
 BX304L01B-BU-00000-r8p0-01rel0
 BX304L06A-BU-00000-r8p0-01rel0
 BX304X07X-BU-00000-r8p0-01rel0

+++ mali_kbase/mali_kbase_jd.c
@@ -1219,7 +1219,7 @@ bool jd_submit_atom(struct kbase_context *kctx, const struct base_jd_atom_v2 *us
 
 #ifdef CONFIG_GPU_TRACEPOINTS
 	katom->work_id = atomic_inc_return(&jctx->work_id);
-	trace_gpu_job_enqueue((u32)kctx->id, katom->work_id,
+	trace_gpu_job_enqueue(kctx->id, katom->work_id,
 			kbasep_map_core_reqs_to_string(katom->core_req));
 #endif
